rename ovis2 to ovis
