+++ vllm/transformers_utils/configs/ovis.py
