+++ vllm/transformers_utils/processors/__init__.py
@@ -2,6 +2,6 @@
 
 from vllm.transformers_utils.processors.deepseek_vl2 import (
     DeepseekVLV2Processor)
-from vllm.transformers_utils.processors.ovis2 import OvisProcessor
+from vllm.transformers_utils.processors.ovis import OvisProcessor
 
 __all__ = ["DeepseekVLV2Processor", "OvisProcessor"]
