+++ vllm/transformers_utils/processors/ovis.py
